+++ mex/backend/extracted/helpers.py
@@ -4,6 +4,8 @@ from pydantic_core import ValidationError
 
 from mex.backend.graph.connector import GraphConnector
 from mex.backend.graph.exceptions import InconsistentGraphError, NoResultFoundError
+from mex.backend.models import ReferenceFilter
+from mex.backend.types import ReferenceFieldName
 from mex.common.models import AnyExtractedModel, PaginatedItemsContainer
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -12,13 +14,11 @@ if TYPE_CHECKING:  # pragma: no cover
     from mex.common.types import Identifier
 
 
-def search_extracted_items_in_graph(  # noqa: PLR0913
+def search_extracted_items_in_graph(
     *,
     query_string: str | None = None,
-    stable_target_id: Identifier | None = None,
     entity_type: Sequence[str] | None = None,
-    referenced_identifiers: Sequence[str] | None = None,
-    reference_field: str | None = None,
+    reference_filters: Sequence[ReferenceFilter] | None = None,
     skip: int = 0,
     limit: int = 100,
 ) -> PaginatedItemsContainer[AnyExtractedModel]:
@@ -26,10 +26,8 @@ def search_extracted_items_in_graph(  # noqa: PLR0913
 
     Args:
         query_string: Full text search query term
-        stable_target_id: Optional stable target ID filter
         entity_type: Optional entity type filter
-        referenced_identifiers: Optional merged item identifiers filter
-        reference_field: Optional field name to filter for
+        reference_filters: Optional reference field filters
         skip: How many items to skip for pagination
         limit: How many items to return at most
 
@@ -43,10 +41,8 @@ def search_extracted_items_in_graph(  # noqa: PLR0913
     graph_result = connector.fetch_extracted_items(
         query_string=query_string,
         identifier=None,
-        stable_target_id=stable_target_id,
         entity_type=entity_type,
-        referenced_identifiers=referenced_identifiers,
-        reference_field=reference_field,
+        reference_filters=reference_filters,
         skip=skip,
         limit=limit,
     )
@@ -77,7 +73,12 @@ def get_extracted_items_from_graph(
         List of extracted items
     """
     search_response = search_extracted_items_in_graph(
-        stable_target_id=stable_target_id,
+        reference_filters=[
+            ReferenceFilter(
+                field=ReferenceFieldName("stableTargetId"),
+                identifiers=[stable_target_id],
+            )
+        ],
         entity_type=entity_type,
         limit=limit,
     )
@@ -90,10 +91,8 @@ def get_extracted_item_from_graph(identifier: Identifier) -> AnyExtractedModel:
     graph_result = connector.fetch_extracted_items(
         query_string=None,
         identifier=str(identifier),
-        stable_target_id=None,
         entity_type=None,
-        referenced_identifiers=None,
-        reference_field=None,
+        reference_filters=None,
         skip=0,
         limit=1,
     )