+++ mex/backend/main.py
@@ -21,7 +21,9 @@ from mex.backend.ldap.main import router as ldap_login_router
 from mex.backend.logging import UVICORN_LOGGING_CONFIG
 from mex.backend.match.main import router as match_router
 from mex.backend.merged.main import router as merged_router
+from mex.backend.migration import migrate
 from mex.backend.preview.main import router as preview_router
+from mex.backend.publishable_merged.main import router as publishable_merged_router
 from mex.backend.rules.main import router as rules_router
 from mex.backend.settings import BackendSettings
 from mex.backend.system.main import router as system_router
@@ -79,6 +81,7 @@ router.include_router(match_router)
 router.include_router(merged_router)
 router.include_router(orcid_router)
 router.include_router(preview_router)
+router.include_router(publishable_merged_router)
 router.include_router(rules_router)
 router.include_router(wikidata_router)
 router.include_router(ldap_login_router)
@@ -108,6 +111,9 @@ def main() -> None:  # pragma: no cover
     on the configured host and port.
     """
     settings = BackendSettings.get()
+
+    migrate()
+
     uvicorn.run(
         "mex.backend.main:app",
         host=settings.backend_host,