+++ mex/backend/security.py
@@ -1,4 +1,4 @@
-from typing import Annotated
+from typing import Annotated, Final
 from urllib.parse import urlsplit
 
 from fastapi import Depends, HTTPException
@@ -12,8 +12,8 @@ from mex.backend.settings import BackendSettings
 from mex.backend.types import APIKey
 from mex.common.logging import logger
 
-X_API_KEY = APIKeyHeader(name="X-API-Key")
-HTTP_BASIC_AUTH = HTTPBasic()
+X_API_KEY: Final = APIKeyHeader(name="X-API-Key")
+HTTP_BASIC_AUTH: Final = HTTPBasic()
 
 
 def has_write_access(